Introduce imagery-source warnings about alignment and reprojections.
Introduce new setting to maps.xsd:
- valid-georeference - which marks imagery sources that are properly georeferenced (i.e. do not need imagery offset adjustments)
- epsg4326to3857Supported - which marks imagery sources that might be safely queried using EPSG:3857 square BBOX, using EPSG:4326 projection

Make the AlignImageryPanel to show only, when valid-georeference is not set to true and provide the ability to hide this message ("do not show again"), which will be now remembered on per-source basis. Due to this changes, good imagery sources will not show warning to novice users and - experienced users will be presented with a warning, when they open an new imagery source which might be unaligned.

Introduce the ability to check "do not show again" the warning about EPSG:3857 to EPSG:4326 reprojections - as above - on per imagery basis. 

All imagery sources should be now reviewed and set valid-georeference=true when appropriate.



git-svn-id: https://josm.openstreetmap.de/svn/trunk@9134 0c6e7542-c601-0410-84e7-c038aed88b3b