Fix thread pool for WMS
Because threds in WMS download ThreadPool never died, the unreferenced ThreadPool object (and its threads) was never cleared. To fix this minimum size of ThreadPool was changed to 0.

But due to the fact, that ThreadPoolExecutor spawns new Thread only when workQueue refuses to take a new job, workaround is implemented in HostLimitQueue based on: http://stackoverflow.com/questions/9622599/java-threadpoolexecutor-strategy-direct-handoff-with-queue#

WMS Settings was fixed so now WMS has it's own THREAD_LIMIT setting not connected to JCS default setting.




git-svn-id: https://josm.openstreetmap.de/svn/trunk@11438 0c6e7542-c601-0410-84e7-c038aed88b3b