Do not cache responses, that are not images.
When server returns content type and it's not an image, then do not put it into
cache and try to extract error message from response.

That way, errors will not be stored in on-disk cache and hopefully they will be
retried when user browses again to the same area (modulo MemoryTileCache).

See: #20443


git-svn-id: https://josm.openstreetmap.de/svn/trunk@17505 0c6e7542-c601-0410-84e7-c038aed88b3b