Fine tune cache sizing
Reduce the cache size, by not counting the tiles that are overloaded for other
zoom levels. Overloaded tiles are only counted for current zoom level.

This reduces memory requirements for a layer, depending on the number of the
tiles shown by 20%-50%. Reduction is high for low number of tiles (when large
tiles or small screen is in use) and low for high number of tiles (when
small tiles or large screen is in use)

See: #20014, #20497



git-svn-id: https://josm.openstreetmap.de/svn/trunk@17516 0c6e7542-c601-0410-84e7-c038aed88b3b